// libs/common/src/errors/error-code.enum.ts
export enum ErrorCode {
  AUTH_MISSING_HEADERS = 'AUTH_001',
  AUTH_INVALID_API_KEY = 'AUTH_002',
  AUTH_API_KEY_INACTIVE = 'AUTH_003',
  AUTH_API_KEY_EXPIRED = 'AUTH_004',
  AUTH_INVALID_SECRET = 'AUTH_005',
  AUTH_CLIENT_APP_INACTIVE = 'AUTH_006',
  AUTH_ADMIN_REQUIRED = 'AUTH_007',
  AUTH_IP_NOT_ALLOWED = 'AUTH_008',
  PERM_INSUFFICIENT = 'PERM_001',
  RATE_LIMIT_EXCEEDED = 'RATE_001',
  MSG_TEMPLATE_NOT_FOUND = 'MSG_001',
  MSG_TEMPLATE_ACCESS_DENIED = 'MSG_002',
  MSG_INVALID_VARIABLES = 'MSG_003',
  MSG_PAYLOAD_NOT_FOUND = 'MSG_004',
  MSG_REQUEST_DATA_MISSING = 'MSG_005',
  MSG_KAFKA_PUBLISH_FAILED = 'MSG_006',
  CLIENT_APP_NOT_FOUND = 'CLIENT_001',
  SYS_INTERNAL_ERROR = 'SYS_001',
}